Uploaded main notebook and LSTM notebook
The LSTM notebook is to code and test the architecture separately before incorportating it into the main notebook. It is inspired by the architecture in https://arxiv.org/pdf/1503.00075.pdf, with some modifications (notably, the weights are phrase type-dependent, and x-bar schemas are treated as trinary nodes, while the other nodes follow a binary tree structure).